Radix counts are longs
Previously if any single bucket surpassed Integer.MAX_VALUE
it would wrap around and become negative, which resulted in
the tracker cache not being fully initialized, which in turn
resulted in an invalid sorting and IdMapper contents as a whole.

+++ community/import-util/src/main/java/org/neo4j/internal/batchimport/cache/idmapping/string/ParallelSort.java
@@ -37,7 +37,7 @@ import static org.neo4j.internal.helpers.Numbers.safeCastLongToInt;
  */
 public class ParallelSort
 {
-    private final int[] radixIndexCount;
+    private final long[] radixIndexCount;
     private final RadixCalculator radixCalculator;
     private final LongArray dataCache;
     private final long highestSetIndex;
